Add files via upload
The code ship_motion_corr_meteor.py contains the coding of the ship motion correction algorithm in the simplified case in which the radar is assumed to be zenith pointing. In this case the processing is simplified. For the algorithm including the complete calculation, please refer to the main branch. This code has been adapted for reading and processing ship data from Meteor and wband radar data postprocessed using the software from Leipzig university. I also included a zip file containing the plots produced with this code on the tested hour. The code is set to run for one hour. To process the entire dataset, it needs to be extended.